requirements.txt
# HTTP and API
requests>=2.28.1
zeep>=4.2.1

# Environment variables
python-dotenv>=0.21.0

# Date and time handling
pytz>=2022.7

# Logging and error handling
structlog>=22.3.0

# Data processing
pandas>=1.5.2

# Caching
diskcache>=5.4.0

# Testing
pytest>=7.2.0
pytest-mock>=3.10.0
